Split runtime configuration helpers
